MYR-376 Route the owner pickup through the dormancy-guarded write
ServePickedUp now calls mutateStatusDispatched, which runs the store's
UpdateStatusFromDispatched. Every other transition keeps the plain guard, and
both variants share one error-mapping and publish path so a refusal or a
success can never be described differently depending on which write produced it.

The refusal reuses the existing `conflict` code — NO new error code — with the
message "a scheduled ride cannot be picked up before its dispatch", so a client
switching on code needs no change and an owner still learns the actual reason
instead of being told their legally-accepted ride has the wrong status.

New telemetry.ErrRideReservationDormant keeps the handler layer decoupled from
internal/store; the cmd adapter translates the store sentinel and funnels both
guarded writes through one guardedUpdate so the row and sentinel translations
cannot drift.

`start` gets no gate and needs none: a dormant reservation can no longer reach
`arrived`, so the existing from-status guard already makes it unstartable.

Two file splits to hold the 300-line rule: the guarded-transition HTTP plumbing
moves to ride_request_status_mutation.go, and the create path — the only
endpoint here with its own precondition chain and error mapper — moves to
ride_request_create_handler.go, taking ride_request_handler.go from 473 lines
to 253. Pure moves; no behaviour travels with them.
