+++ internal/telemetry/ride_request_handler.go
@@ -85,171 +85,6 @@ func NewRideRequestHandler(
 	return h
 }
 
-// ServeCreate handles POST /api/ride-requests.
-func (h *RideRequestHandler) ServeCreate(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	userID, ok := h.authUser(w, r)
-	if !ok {
-		return
-	}
-
-	body, ok := h.decodeCreateBody(w, r)
-	if !ok {
-		return
-	}
-
-	in, reason := validateCreateBody(body)
-	if reason != "" {
-		h.writeError(w, http.StatusBadRequest, wserrors.ErrCodeInvalidRequest, reason)
-		return
-	}
-
-	// Derive ownerId from the vehicle + enforce vehicle access: the owner,
-	// or a viewer at the `rides` tier (see the type doc).
-	row, err := h.vehicles.GetByID(ctx, in.VehicleID)
-	if err != nil {
-		if errors.Is(err, sdk.ErrNotFound) {
-			h.writeError(w, http.StatusNotFound, wserrors.ErrCodeNotFound, "vehicle not found")
-			return
-		}
-		h.logger.Error("ride-request create: vehicle lookup failed",
-			slog.String("vehicle_id", in.VehicleID),
-			slog.String("error", err.Error()),
-		)
-		h.writeError(w, http.StatusInternalServerError, wserrors.ErrCodeInternalError, "internal error")
-		return
-	}
-	// MYR-369: the gate reads the grant's `allow_rides` FLAG, not a tier. A
-	// viewer whose grant does not carry it is refused here even though they
-	// can see the car — and the owner can now take the capability away in
-	// place (PATCH /api/invites/{inviteId}) instead of revoking the whole
-	// grant, so this check is live state rather than something fixed at
-	// invite time. capRides carries the suspension term itself, so a
-	// suspended grant is refused here as well as being absent from the
-	// access set.
-	if _, accessErr := vehicleAccessFor(ctx, h.shares, userID, in.VehicleID, row.UserID, capRides); accessErr != nil {
-		if errors.Is(accessErr, errNoVehicleAccess) {
-			denyVehicleAccess(w, h.logger, "ride-request create", in.VehicleID, userID)
-			return
-		}
-		h.logger.Error("ride-request create: access resolution failed",
-			slog.String("vehicle_id", in.VehicleID),
-			slog.String("error", accessErr.Error()),
-		)
-		h.writeError(w, http.StatusInternalServerError, wserrors.ErrCodeInternalError, "internal error")
-		return
-	}
-	// MYR-342: the vehicle's owner may have PAUSED ride requests for this car.
-	// Placed immediately after the access check and BEFORE anything is derived
-	// or written, so a caller with no access to the vehicle still gets the
-	// access denial rather than learning its pause state (ride_share_gate.go).
-	//
-	// Applies to INSTANT AND SCHEDULED creates alike — a deliberate deviation
-	// from the MYR-313 scheduled exemption below it. A service visit ends on its
-	// own, so a reservation booked across one is likely servable; an owner's
-	// pause is open-ended, so a reservation booked against a paused car is not.
-	// The flag is read off `row`, the very lookup that resolved the owner, so
-	// access and availability come from one statement and cannot disagree.
-	if rejectIfRideSharePaused(w, h.logger, "ride-request create", row, userID) {
-		return
-	}
-
-	in.RiderID = userID
-	// The ride's owner is the VEHICLE's owner, which for a shared rider is
-	// somebody other than the requester — that asymmetry is the point of the
-	// `rides` tier, and it is what routes the accept/decline to the right
-	// person.
-	in.OwnerID = row.UserID
-
-	// MYR-316: a scheduled ride may not be booked for a time before the car is
-	// expected back from service. Instant rides are unaffected (already gated
-	// by MYR-277), and a vehicle with no estimate imposes no bound.
-	if h.rejectIfBeforeServiceWindow(w, in.ScheduledFor, row) {
-		return
-	}
-
-	// One active ride per rider (MYR-230): an INSTANT request is refused
-	// while the rider already has an open instant ride. The fast-path
-	// pre-check runs here; the partial unique index (migration 0004) is the
-	// authoritative race backstop, applied to the Create error below.
-	if h.rejectIfRideActive(ctx, w, in) {
-		return
-	}
-
-	created, err := h.store.Create(ctx, in)
-	if err != nil {
-		h.handleCreateError(ctx, w, in, err)
-		return
-	}
-
-	h.publish(ctx, events.RideRequestCreatedEvent{
-		RideRequestID: created.ID,
-		VehicleID:     created.VehicleID,
-		RiderID:       created.RiderID,
-		OwnerID:       created.OwnerID,
-		Status:        created.Status,
-		RequesterName: created.RequesterName,
-		ScheduledFor:  created.ScheduledFor,
-		CreatedAt:     created.CreatedAt,
-	})
-
-	h.writeJSON(w, http.StatusCreated, toRideRequestWire(created))
-}
-
-// rejectIfRideActive is the one-active-instant-ride fast path (MYR-230). For
-// an INSTANT request it 409s `ride_active` (carrying the existing ride to
-// adopt) when the rider already has an open instant ride, or 500s on a lookup
-// failure — returning true in either case so ServeCreate stops. Scheduled
-// requests are EXEMPT (a future reservation is not an active ride) and always
-// pass through (returns false). The partial unique index stays the
-// authoritative race guard, enforced by handleCreateError.
-func (h *RideRequestHandler) rejectIfRideActive(ctx context.Context, w http.ResponseWriter, in RideRequestCreateInput) bool {
-	if in.ScheduledFor != nil {
-		return false
-	}
-	switch existing, err := h.store.GetActiveInstantByRider(ctx, in.RiderID); {
-	case err == nil:
-		h.writeRideActive(w, existing)
-		return true
-	case errors.Is(err, sdk.ErrNotFound):
-		return false // no open instant ride — proceed to create
-	default:
-		h.logger.Error("ride-request create: active-ride pre-check failed",
-			slog.String("user_id", in.RiderID),
-			slog.String("error", err.Error()),
-		)
-		h.writeError(w, http.StatusInternalServerError, wserrors.ErrCodeInternalError, "internal error")
-		return true
-	}
-}
-
-// handleCreateError maps a failed store.Create onto a response. ErrRideActive
-// (the unique-index race backstop refusing a concurrent second instant
-// create) re-reads the winning ride so the client adopts it, exactly as the
-// pre-check would; every other error is a 500.
-func (h *RideRequestHandler) handleCreateError(ctx context.Context, w http.ResponseWriter, in RideRequestCreateInput, err error) {
-	if errors.Is(err, ErrRideActive) {
-		if existing, gErr := h.store.GetActiveInstantByRider(ctx, in.RiderID); gErr == nil {
-			h.writeRideActive(w, existing)
-			return
-		}
-		// Extreme TOCTOU: the conflicting ride reached a terminal state
-		// between the rejected insert and this re-read. Surface the typed 409
-		// without a body rather than a generic failure; the SDK re-syncs its
-		// own list and may retry the create.
-		h.logger.Warn("ride-request create: active guard fired but open ride vanished",
-			slog.String("user_id", in.RiderID),
-		)
-		h.writeError(w, http.StatusConflict, wserrors.ErrCodeRideActive, "you already have an active ride request")
-		return
-	}
-	h.logger.Error("ride-request create: store failed",
-		slog.String("vehicle_id", in.VehicleID),
-		slog.String("error", err.Error()),
-	)
-	h.writeError(w, http.StatusInternalServerError, wserrors.ErrCodeInternalError, "internal error")
-}
-
 // ServeCancel handles POST /api/ride-requests/{id}/cancel. Rider-only;
 // legal only from requested/accepted (→ cancelled). Every other current
 // status is a 409 conflict.
@@ -289,63 +124,8 @@ func (h *RideRequestHandler) ServeCancel(w http.ResponseWriter, r *http.Request)
 // in lockstep with cancellableFrom and the rest-api.md §7.8 matrix.
 var rideCancellableFrom = []string{rideStatusRequested, rideStatusAccepted}
 
-// mutateStatus persists a status change ATOMICALLY via the store's guarded
-// UpdateStatusFrom — a single UPDATE with `WHERE status = ANY(from)` — so a
-// check-then-write race (rider-cancel vs owner-decline, owner double-accept)
-// resolves in the database: exactly one caller wins; the loser gets a 409
-// conflict here even if its pre-check read passed. Only the winning call
-// publishes the ride_status_changed summary event (and, for accept, the
-// ride.accepted dispatch seam at the call site) — exactly-once by
-// construction. Does NOT write the success response; the caller does. On
-// failure writes the error envelope and returns ok=false. Shared by cancel
-// (MYR-174) and accept/decline (MYR-175).
-func (h *RideRequestHandler) mutateStatus(ctx context.Context, w http.ResponseWriter, rec RideRequestData, from []string, to string) (RideRequestData, bool) {
-	updated, err := h.store.UpdateStatusFrom(ctx, rec.ID, from, to)
-	if err != nil {
-		switch {
-		case errors.Is(err, ErrVehicleRideActive):
-			// Per-vehicle one-active-ride guard (MYR-266): the transition is
-			// legal for THIS ride, but the target vehicle is already committed
-			// to another active ride — a capability gate on the vehicle, the
-			// same 409 vehicle_unavailable class as the MYR-277 in-service /
-			// offline gate (the car cannot serve a second ride at once), NOT the
-			// illegal-transition `conflict`.
-			h.writeError(w, http.StatusConflict, wserrors.ErrCodeVehicleUnavailable, "Vehicle is already on another ride")
-		case errors.Is(err, ErrRideStatusConflict):
-			// Lost the transition race (or the pre-check read was stale):
-			// the row's current status left the allowed-from set between
-			// our read and the guarded write.
-			h.writeError(w, http.StatusConflict, wserrors.ErrCodeConflict, "ride request cannot transition to "+to+" from its current status")
-		case errors.Is(err, sdk.ErrNotFound):
-			// Raced with a delete between GetByID and the guarded write.
-			h.writeError(w, http.StatusNotFound, wserrors.ErrCodeNotFound, "ride request not found")
-		default:
-			h.logger.Error("ride-request transition: store failed",
-				slog.String("ride_request_id", rec.ID),
-				slog.String("status", to),
-				slog.String("error", err.Error()),
-			)
-			h.writeError(w, http.StatusInternalServerError, wserrors.ErrCodeInternalError, "internal error")
-		}
-		return RideRequestData{}, false
-	}
-
-	h.publish(ctx, events.RideStatusChangedEvent{
-		RideRequestID:    updated.ID,
-		VehicleID:        updated.VehicleID,
-		RiderID:          updated.RiderID,
-		OwnerID:          updated.OwnerID,
-		Status:           updated.Status,
-		RequesterName:    updated.RequesterName,
-		RescheduleStatus: updated.RescheduleStatus,
-		// Carried so the rider's push copy can tell a reservation from an
-		// instant request (MYR-360); not projected onto the WS frame.
-		ScheduledFor: updated.ScheduledFor,
-		UpdatedAt:    updated.UpdatedAt,
-	})
-
-	return updated, true
-}
+// The guarded-transition helpers (mutateStatus and its dormancy-guarded
+// sibling) live in ride_request_status_mutation.go.
 
 // cancellableFrom reports whether a rider cancel is legal from the given
 // status. Legal only from requested/accepted; enroute/arrived (ride in
