+++ internal/telemetry/ride_request_owner_progress_handler.go
@@ -29,6 +29,27 @@ import (
 // ServePickedUp handles POST /api/ride-requests/{id}/picked-up. Owner-only;
 // legal only from accepted → arrived (409 otherwise). Idempotent: an already
 // `arrived` ride is a 200 no-op that re-returns the current record.
+//
+// RESERVATION DORMANCY (MYR-376). A pickup is additionally refused — 409, the
+// SAME `conflict` code, a message naming the reason — when the ride is a
+// SCHEDULED one that is neither dispatched nor yet due (`scheduledFor` set and
+// in the FUTURE, and dispatch not resolved `sent`). A reservation is DORMANT
+// from accept until the earlier of its dispatch and its due instant: the car
+// has been told nothing, so there is no pickup to confirm. In production an
+// owner could tap "Picked up" the day before the ride was due and strand it at
+// `arrived` with no legal exit.
+//
+// Dormancy ENDS AT THE DUE INSTANT, dispatched or not — a reservation the
+// sweeper failed (`reservation_expired`), skipped, or never reached is picked
+// up normally once `scheduledFor` passes, which is what makes §7.8's "may
+// still cancel or proceed manually" promise real rather than a dead end. The
+// only thing refused is a pickup BEFORE the ride was due.
+//
+// The refusal is enforced INSIDE the guarded write (mutateStatusDispatched),
+// never as a pre-check, so it is race-correct against a concurrently
+// dispatching sweeper and against the clock crossing `scheduledFor`. INSTANT
+// rides are entirely unaffected, including ones whose nav push failed or was
+// skipped.
 func (h *RideRequestHandler) ServePickedUp(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	userID, ok := h.authUser(w, r)
@@ -52,7 +73,9 @@ func (h *RideRequestHandler) ServePickedUp(w http.ResponseWriter, r *http.Reques
 		return
 	}
 
-	updated, ok := h.mutateStatus(ctx, w, rec, ridePickedUpFrom, rideStatusArrived)
+	// The DORMANCY-guarded write (MYR-376): the reservation predicate rides the
+	// same UPDATE as the status guard, so neither refusal can be raced past.
+	updated, ok := h.mutateStatusDispatched(ctx, w, rec, ridePickedUpFrom, rideStatusArrived)
 	if !ok {
 		return
 	}
@@ -94,8 +117,9 @@ func (h *RideRequestHandler) ServeDroppedOff(w http.ResponseWriter, r *http.Requ
 
 // ridePickedUpFrom / rideDroppedOffFrom are the owner handshake allowed-from
 // sets; must stay in lockstep with the fast-path checks above and the
-// rest-api.md §7.8 matrix. The guarded UpdateStatusFrom write is what decides
-// under concurrency.
+// rest-api.md §7.8 matrix. The guarded write is what decides under concurrency
+// — and for pickup that write also carries the MYR-376 dormancy predicate, so
+// the allowed-from set is not the whole precondition.
 var (
 	ridePickedUpFrom   = []string{rideStatusAccepted}
 	rideDroppedOffFrom = []string{rideStatusEnroute}
